[Misc] Added link between 'room' and 'user' via field 'creator'

+++ src/handlers/room/room.delete.ts
@@ -5,6 +5,7 @@ import { Connection } from 'typeorm';
 import Room from '../../models/mysql/Room';
 import log from '../../logger'
 import Packet from '../../packet';
+import User from '../../models/mysql/User';
 
 export class RoomDelete extends Handler {
     public deleteById(req: Request, res: Response, next: NextFunction): void {
@@ -43,15 +44,19 @@ export class RoomDelete extends Handler {
         connect.then(async connection => {
             let packet = new Packet('room', 'deleteAllByCreatorId');
             if (req.query.creator_id) {
-                let creator_id: number = Number(req.query.creator_id);
+                let creator_id: number = Number(req.query.creator);
                 if (!isNaN(creator_id)) {
                     if (connection instanceof Connection && connection.isConnected) {
-                        let roomRepository = connection.getRepository(Room);
-                        let data = await roomRepository.find({creator_id: creator_id});
-                        if (!data || data.length == 0) {
-                            packet.error = `No rooms with creator_id ${creator_id}`;
+                        let userRepository = connection.getRepository(User);
+                        let creator: User = await userRepository.findOneById(creator_id);
+                        if (creator) {
+                            let roomRepository = connection.getRepository(Room);
+                            let data = await roomRepository.find({creator: creator});
+                            if (!data || data.length == 0) {
+                                packet.error = `No rooms with creator_id ${creator_id}`;
+                            }
+                            await roomRepository.remove(data);
                         }
-                        await roomRepository.remove(data);
                     } else {
                         log.error('typeorm', 'DBConnection error');
                         packet.error = 'Internal error';
